More refactoring + more comments
