+++ source/LottieData/LottieData.projitems
@@ -37,7 +37,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)NullLayer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\LayerGroup.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\LayersGraph.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\ShiftUtils.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\MergeHelper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\ShiftHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Optimization\GradientStopOptimizer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\LayersIndexMapper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Optimization\Experimental\LottieCompositionOptimizer.cs" />
