+++ source/LottieData/Optimization/Experimental/LayersIndexMapper.cs
@@ -2,6 +2,13 @@
 
 namespace Microsoft.Toolkit.Uwp.UI.Lottie.LottieData.Optimization
 {
+    /// <summary>
+    /// Helper data structure that manages layers index remapping.
+    /// </summary>
+#if PUBLIC_LottieData
+    public
+#endif
+
     class LayersIndexMapper
     {
         Dictionary<int, int> indexMapping = new Dictionary<int, int>();
@@ -18,9 +25,11 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie.LottieData.Optimization
 
         public void RemapLayer(Layer layer)
         {
+            // TODO: create layer copy instad of assigning new index(!)
             layer.Index = GetMapping(layer.Index);
             if (layer.Parent is not null)
             {
+                // TODO: same as above(!)
                 layer.Parent = GetMapping((int)layer.Parent);
             }
         }
